# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.APIResponse:
    description: Standard API response format with metadata and context information
    properties:
      data:
        description: Response data (varies by endpoint)
      error:
        allOf:
        - $ref: '#/definitions/dto.ErrorResponse'
        description: Error details (only present when success=false)
      meta:
        allOf:
        - $ref: '#/definitions/dto.ResponseMetadata'
        description: Response metadata with request tracking
      pagination:
        allOf:
        - $ref: '#/definitions/dto.PaginationMetadata'
        description: Pagination metadata (for paginated responses)
      success:
        description: Indicates if the request was successful
        example: true
        type: boolean
    type: object
...
  dto.ErrorResponse:
    description: Error response structure with detailed error information
    properties:
      code:
        description: Error code for programmatic handling
        example: VALIDATION_ERROR
        type: string
      details:
        additionalProperties: true
        description: Additional error details
        type: object
      message:
        description: Human-readable error message
        example: Invalid input data
        type: string
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Error timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
    type: object
...
  dto.PaginationMetadata:
    description: Pagination metadata for paginated responses
    properties:
      has_next:
        description: Whether there is a next page
        example: true
        type: boolean
      has_prev:
        description: Whether there is a previous page
        example: false
        type: boolean
      limit:
        description: Number of items per page
        example: 10
        type: integer
      page:
        description: Current page number (1-based)
        example: 1
        type: integer
      request_id:
        description: Request ID for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      total:
        description: Total number of items
        example: 100
        type: integer
      total_pages:
        description: Total number of pages
        example: 10
        type: integer
    type: object
  dto.ResponseMetadata:
    description: Response metadata containing request tracking and performance information
    properties:
      processing_time_ms:
        description: Request processing time in nanoseconds
        example: 15000000
        type: integer
      request_id:
        description: Unique request identifier for tracking
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        description: Response generation timestamp
        example: "2025-07-24T20:43:16.123456789+07:00"
        type: string
      version:
        description: API version
        example: 1.0.0
        type: string
    type: object
...
  model.Article:
    properties:
      category:
        $ref: '#/definitions/model.Category'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      slug:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/model.User'
      views:
        type: integer
    type: object
...
  model.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  model.Likes:
    properties:
      article:
        $ref: '#/definitions/model.Article'
      created_at:
        type: string
      id:
        type: integer
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4300
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk aplikasi blog Develapar dengan fitur lengkap untuk manajemen
    artikel, komentar, kategori, tag, bookmark, dan like. API menggunakan standard
    response format dengan metadata, request tracking, rate limiting, dan comprehensive
    error handling.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Develapar API
  version: "1.0"
paths:
...
  /likes:
    delete:
      consumes:
      - application/json
      description: Remove a like from a specific article by the authenticated user
      parameters:
      - description: Article ID to unlike
        in: body
        name: payload
        required: true
        schema:
          properties:
            article_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Like deleted successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid article ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Remove a like from an article
      tags:
      - Likes
    post:
      consumes:
      - application/json
      description: Add a like to a specific article by the authenticated user
      parameters:
      - description: Like creation details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Likes'
      produces:
      - application/json
      responses:
        "201":
          description: Like successfully added
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    like:
                      $ref: '#/definitions/model.Likes'
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid payload
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Add a like to an article
      tags:
      - Likes
  /likes/article/{article_id}:
    get:
      description: Get a list of likes for a specific article ID
      parameters:
      - description: ID of the article to retrieve likes for
        in: path
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of likes for the article
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    likes:
                      items:
                        $ref: '#/definitions/model.Likes'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid article ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get likes by article ID
      tags:
      - Likes
  /likes/check:
    get:
      description: Check if a specific article is liked by the authenticated user
      parameters:
      - description: ID of the article to check
        in: query
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Like status
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    liked:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Invalid article ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Check if an article is liked by the current user
      tags:
      - Likes
  /likes/user/{user_id}:
    get:
      description: Get a list of likes by a specific user ID
      parameters:
      - description: ID of the user whose likes to retrieve
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of likes by the user
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    likes:
                      items:
                        $ref: '#/definitions/model.Likes'
                      type: array
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid user ID
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "408":
          description: Request timeout
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/dto.ErrorResponse'
              type: object
      summary: Get likes by user ID
      tags:
      - Likes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"